Complete CRUD

+++ ProductInventory.Api/src/Repositories/ProductRepository.cs
@@ -1,51 +1,64 @@
+using System;
 using Microsoft.AspNetCore.Http.HttpResults;
+using Microsoft.EntityFrameworkCore;
 using ProductInventory.Api.Data;
 using ProductInventory.Api.Models;
 using ProductInventory.Api.Service;
 
 namespace ProductInventory.Api.Repositories;
 
-public class ProductRepositories : IProductRepository
+public class ProductRepository : IProductRepository
 {
-    // List<Products> products;
-    public ApplicationDbContext _context;
-    public ProductRepositories(ApplicationDbContext applicationDbContext)
+    private readonly ApplicationDbContext _context;
+
+    public ProductRepository(ApplicationDbContext context)
     {
-        _context = applicationDbContext;
-    }
-    public Products Get(string id)
-    {
-        Products products = _context.products.Find(id);
-        return products;
+        _context = context;
     }
 
-   public List<Products> GetAll()
-{
-    return _context.products.ToList<Products>();
-}
-
-
-    public Products RemoveProducts(string id)
+    public async Task<Products> AddAsync(Products product)
     {
-        var product = Get(id);
+        await _context.products.AddAsync(product);
+        await _context.SaveChangesAsync();
+        return product;
+    }
+
+    public async Task<IEnumerable<Products>> GetProductsAsync()
+    {
+        return await _context.products.ToListAsync();
+    }
+
+    public async Task DeleteAsync(Guid id)
+    {
+        var product = await _context.products.FindAsync(id);
+        if (product is null)
+        {
+            return;
+        }
+
         _context.products.Remove(product);
-        _context.SaveChanges();
-        return product;
-
+        await _context.SaveChangesAsync();
+        return;
     }
 
-    public Products Save(Products product)
+
+
+    public async Task<Products> GetByIdAsync(Guid id)
     {
-        _context.products.Add(product);
-        _context.SaveChanges();
-        return product;
+        return await _context.products.FindAsync(id)!;
+    }
+    
+
+    public async Task UpdateAsync(Products product)
+    {
+        var existingProduct = await _context.products.FindAsync(product.Id);
+        if (existingProduct is null)
+        {
+            throw new KeyNotFoundException("Product not found");
+        }
+
+        _context.Entry(existingProduct).CurrentValues.SetValues(product);
+        await _context.SaveChangesAsync();
     }
 
-    public Products UpdateProduct(Products products)
-    {
-        _context.Entry(products).CurrentValues.SetValues(products);
-        _context.SaveChanges();
-        return products;
-        
-    }
 }
